Extract CompatibilityTestCommon
